backend/requirements.txt
supabase==2.4.0
python-dotenv==1.0.1
argparse==1.4.0
